requirements.txt
PySide6>=6.6
opencv-python-headless>=4.9
pyserial>=3.5
numpy>=1.24
matplotlib>=3.8
Pillow>=10.0